% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censusDataTasks.R
\name{concatCensusGroup}
\alias{concatCensusGroup}
\title{Dataset by census tract for whole US}
\usage{
concatCensusGroup(
  name = "acs/acs5",
  start_year = 2009,
  end_year = 2018,
  tablename = "age_data",
  group = "B01001",
  region = "tract:*",
  fipstate = fips
)
}
\arguments{
\item{name}{(character) the name of the census api dataset}

\item{start_year}{(numeric) first year of longitudal data}

\item{end_year}{(numeric) last year of longitudinal data}

\item{tablename}{(character) name to assign dataframe generated by function}

\item{group}{(character) the name of the metadata group}
}
\description{
Dataset by census tract for whole US
}
