% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censusDataTasks.R
\name{concatCensusVars}
\alias{concatCensusVars}
\title{Age demographics by census tract for whole US}
\usage{
concatCensusVars(
  name = "zbp",
  start_year = 2009,
  end_year = 2018,
  tablename = "zip_code_bp",
  region = "tract:*"
)
}
\arguments{
\item{name}{(character) the name of the census api dataset}

\item{start_year}{(numeric) first year of longitudal data}

\item{end_year}{(numeric) last year of longitudinal data}

\item{tablename}{(character) name to assign dataframe generated by function}
}
\description{
Age demographics by census tract for whole US
}
