% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censusDataTasks.R
\name{getCensusGroups}
\alias{getCensusGroups}
\title{View the available groups for each census api endpoint}
\usage{
getCensusGroups(name = "acs/acs5", year = 2018)
}
\arguments{
\item{name}{(character) name of the census api enpoint}

\item{year}{(numeric) vintage of the dataset}
}
\description{
View the available groups for each census api endpoint
}
