% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censusDataTasks.R
\name{getCensusInfo}
\alias{getCensusInfo}
\title{Get variable and geography metadata in a single function}
\usage{
getCensusInfo(name = "acs/acs5", year = 2018)
}
\arguments{
\item{name}{(character) name of censusapi dataset}

\item{year}{(numeric) vintage year}
}
\description{
Get variable and geography metadata in a single function
}
