% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredDataTasks.R
\name{getFred}
\alias{getFred}
\title{Download a FRED dataset and store it as a dataframe}
\usage{
getFred(d_id, fred_df = "fred_data")
}
\arguments{
\item{d_id}{(character) the FRED dataset ID}

\item{fred_df}{(character) the desired name of the FRED data dataframe}
}
\description{
Download a FRED dataset and store it as a dataframe
}
