% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredDataTasks.R
\name{searchFred}
\alias{searchFred}
\title{Get Look from Looker instance without row limits}
\usage{
searchFred(key, search_df = "results")
}
\arguments{
\item{key}{(character) the search term to lookup FRED datasets}

\item{search_df}{(character) the desired name of the search results dataframe}
}
\description{
Get Look from Looker instance without row limits
}
